+++ src/main/java/com/ubatis/circleserver/bean/LogCircleLoginBean.java
@@ -8,9 +8,9 @@ public class LogCircleLoginBean implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /** id */ 
-    private int id;
+    private long id;
     /** 登录的圈子账号id */ 
-    private int circle_id;
+    private long circle_id;
     /**  */ 
     private String ip_address;
     /**  */ 
@@ -19,10 +19,10 @@ public class LogCircleLoginBean implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
-    public int getCircle_id() { 
+    public long getCircle_id() { 
         return circle_id;
     }
     public String getIp_address() { 
@@ -36,10 +36,10 @@ public class LogCircleLoginBean implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         this.id = id;
     }
-    public void setCircle_id(int circle_id) { 
+    public void setCircle_id(long circle_id) { 
         this.circle_id = circle_id;
     }
     public void setIp_address(String ip_address) {
