+++ src/main/java/com/ubatis/circleserver/bean/LogUserLoginBean.java
@@ -8,9 +8,9 @@ public class LogUserLoginBean implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /**  */ 
-    private int id;
+    private long id;
     /** 登录id */ 
-    private int user_id;
+    private long user_id;
     /**  */ 
     private String ip_address;
     /**  */ 
@@ -19,10 +19,10 @@ public class LogUserLoginBean implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
-    public int getUser_id() { 
+    public long getUser_id() { 
         return user_id;
     }
     public String getIp_address() { 
@@ -36,10 +36,10 @@ public class LogUserLoginBean implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         this.id = id;
     }
-    public void setUser_id(int user_id) { 
+    public void setUser_id(long user_id) { 
         this.user_id = user_id;
     }
     public void setIp_address(String ip_address) {
