+++ src/main/java/com/ubatis/circleserver/bean/LogUserOperationBean.java
@@ -8,9 +8,9 @@ public class LogUserOperationBean implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /**  */ 
-    private int id;
+    private long id;
     /**  */ 
-    private int user_id;
+    private long user_id;
     /**  */ 
     private String operate_name;
     /**  */ 
@@ -21,10 +21,10 @@ public class LogUserOperationBean implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
-    public int getUser_id() { 
+    public long getUser_id() { 
         return user_id;
     }
     public String getOperate_name() { 
@@ -41,10 +41,10 @@ public class LogUserOperationBean implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         this.id = id;
     }
-    public void setUser_id(int user_id) { 
+    public void setUser_id(long user_id) { 
         this.user_id = user_id;
     }
     public void setOperate_name(String operate_name) {
