+++ src/main/java/com/ubatis/circleserver/bean/param/ParamAcUser.java
@@ -9,9 +9,9 @@ public class ParamAcUser extends MyParams implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /** id */ 
-    private int id;
+    private long id;
     /**  */ 
-    private int circle_id;
+    private long circle_id;
     /**  */ 
     private String openid;
     /**  */ 
@@ -38,10 +38,10 @@ public class ParamAcUser extends MyParams implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
-    public int getCircle_id() { 
+    public long getCircle_id() { 
         return circle_id;
     }
     public String getOpenid() { 
@@ -82,11 +82,11 @@ public class ParamAcUser extends MyParams implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         put("id",id);
         this.id = id;
     }
-    public void setCircle_id(int circle_id) { 
+    public void setCircle_id(long circle_id) { 
         put("circle_id",circle_id);
         this.circle_id = circle_id;
     }
