+++ src/main/java/com/ubatis/circleserver/bean/param/ParamLogCircleLogin.java
@@ -9,9 +9,9 @@ public class ParamLogCircleLogin extends MyParams implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /** id */ 
-    private int id;
+    private long id;
     /** 登录的圈子账号id */ 
-    private int circle_id;
+    private long circle_id;
     /**  */ 
     private String ip_address;
     /**  */ 
@@ -20,10 +20,10 @@ public class ParamLogCircleLogin extends MyParams implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
-    public int getCircle_id() { 
+    public long getCircle_id() { 
         return circle_id;
     }
     public String getIp_address() { 
@@ -37,11 +37,11 @@ public class ParamLogCircleLogin extends MyParams implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         put("id",id);
         this.id = id;
     }
-    public void setCircle_id(int circle_id) { 
+    public void setCircle_id(long circle_id) { 
         put("circle_id",circle_id);
         this.circle_id = circle_id;
     }
