+++ src/main/java/com/ubatis/circleserver/bean/param/ParamLogUserLogin.java
@@ -9,9 +9,9 @@ public class ParamLogUserLogin extends MyParams implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /**  */ 
-    private int id;
+    private long id;
     /** 登录id */ 
-    private int user_id;
+    private long user_id;
     /**  */ 
     private String ip_address;
     /**  */ 
@@ -20,10 +20,10 @@ public class ParamLogUserLogin extends MyParams implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
-    public int getUser_id() { 
+    public long getUser_id() { 
         return user_id;
     }
     public String getIp_address() { 
@@ -37,11 +37,11 @@ public class ParamLogUserLogin extends MyParams implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         put("id",id);
         this.id = id;
     }
-    public void setUser_id(int user_id) { 
+    public void setUser_id(long user_id) { 
         put("user_id",user_id);
         this.user_id = user_id;
     }
