+++ src/main/java/com/ubatis/circleserver/bean/param/ParamSysDict.java
@@ -9,7 +9,7 @@ public class ParamSysDict extends MyParams implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /**  */ 
-    private int id;
+    private long id;
     /**  */ 
     private String category;
     /**  */ 
@@ -28,7 +28,7 @@ public class ParamSysDict extends MyParams implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
     public String getCategory() { 
@@ -57,7 +57,7 @@ public class ParamSysDict extends MyParams implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         put("id",id);
         this.id = id;
     }
