+++ src/main/java/com/ubatis/circleserver/bean/param/ParamSysUserFormid.java
@@ -9,9 +9,9 @@ public class ParamSysUserFormid extends MyParams implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /**  */ 
-    private int id;
+    private long id;
     /**  */ 
-    private int user_id;
+    private long user_id;
     /**  */ 
     private String form_id;
     /**  */ 
@@ -20,10 +20,10 @@ public class ParamSysUserFormid extends MyParams implements Serializable {
     private String gmt_update;
 
     //getter
-    public int getId() { 
+    public long getId() { 
         return id;
     }
-    public int getUser_id() { 
+    public long getUser_id() { 
         return user_id;
     }
     public String getForm_id() { 
@@ -37,11 +37,11 @@ public class ParamSysUserFormid extends MyParams implements Serializable {
     }
 
     //setter
-    public void setId(int id) { 
+    public void setId(long id) { 
         put("id",id);
         this.id = id;
     }
-    public void setUser_id(int user_id) { 
+    public void setUser_id(long user_id) { 
         put("user_id",user_id);
         this.user_id = user_id;
     }
